Fix example patch tests and stop tracking prev and next tokens (#15)
The example patch tests were broken because the parser was
initialized but never run. Additionally, none of the example patches
were complex enough to hit the recursion depth limit. I added a few
longer patches and run the parser, which reproduced the recursion
depth limit within tests.

To fix the recursion limit, which was caused by deep copying tokens,
and by extension all other tokens given that they are doubly linked,
I removed tracking of previous and next tokens. This was initially
implemented to handle multi-word instructions by traversing to
neighboring tokens, but that is now handled by TokenRegistry, and
there are no other apparent use cases. If I find that I need that
functionality in the future, it can be re-implemented and the _clone
method can be rewritten to avoid deeply copying neighboring tokens.